src/ClubKitApi/requirements.txt
dj-database-url==0.5.0

dj-static==0.0.6

Django==2.0.2

django-cors-middleware==1.3.1

djangorestframework==3.7.7

python-decouple==3.1

pytz==2018.3

static3==0.7.0
